packages/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
orjson==3.10.7
PyYAML==6.0.2
